feat: Add DeckMaker for makeCards
